requirements.txt
streamlit==1.30.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
requests==2.31.0
appbuilder-sdk==0.3.4
numpy==1.24.3